able to extract the company profile detail
