+++ extractor/src/plugins/bursa-malaysia/extract/equity_page/spec/mock/equities_prices.html
